random point subtraction

+++ catalog-validations-code/mdet_systematics.py
@@ -447,8 +447,8 @@ def tangential_shear_field_center(fs):
                 d = fio.read(cat2_f)
                 mask_noshear = (d['mdet_step'] == 'noshear')
                 print(d[mask_noshear]['mdet_g_1'])
-                g1 = d[mask_noshear]['mdet_g_1']/R11
-                g2 = d[mask_noshear]['mdet_g_2']/R22
+                g1 = d[mask_noshear]['mdet_g_1']/np.float64(R11)
+                g2 = d[mask_noshear]['mdet_g_2']/np.float64(R22)
                 cat2 = treecorr.Catalog(ra=d['ra'], dec=d['dec'], ra_units='deg', dec_units='deg', g1=g1, g2=g2, patch_centers=cat1.patch_centers)
             
                 ng_rand.process(cat1, cat2, initialize=(i==0), finalize=(i==len(cat2_files)-1))
